Added a keyhole animation
I added a key insertion and turning animation, added icons for the keyhole and key items, and made a short script that starts the animation and removes the keyhole after. I don't know if this script is necessary, but it was the easiest way I could think of because you can add an event to an animation to make a function trigger, so I made a function to hide the keyhole.