Starting something
Can you make it so that the three bars on the handle are missing (so you can't open it at first) and you have to take the y in "play" to add them so you can spin the wheel and open the door? I also added the room for the office area.