A lot of changes
Organization, implemented the Y, and improved the movement controller.

BUGS/MISSING FEATURES SO FAR THAT I HAVE FOUND: Keyhole animation breaks when you press play (maybe we can make it so you can't press play when the keyhole animation is running), movement controller is still pretty bad, hole item on canvas goes out of frame, no Y item texture (using brick as placeholder), no play button without Y texture, no Y texture to go on top of that texture (using brick as placeholder) (note: I have a solution for the no y on door texture), and maybe we could make the parts of the play button falling a little more random with animation which also would make the hole not go offscreen (and they also fall differently if you pause while they are falling).

Also Adrian had an idea of "skeletons in the closet" for like a bedroom or something with a closet but one of the bones of the skeleton is a "skeleton key" which opens a door (but it gets used up). I also thought that maybe we could do hint messages appearing if you spend a while on a puzzle (that can be turned on and are off by default or maybe the opposite of that). For example:

A while on pause puzzle: "Think outside of the box"
Even longer on pause puzzle (it will have already told you the controls earlier: "Remember: Esc to pause"
Even longer on pause puzzle: "Esc to go outside of the box"
Way too long on pause puzzle (for other ones would also just give the solution): "Go into the pause menu and try to move around after a second to complete the puzzle"

I think either that or just a hint section of the menu would be good.

Also Adrian played through it and said that the Y puzzle was really hard because of all of the different symbols in the logo so we should definitely add hints for that.

Sorry this is so long but one last thing I think there could be like a piece of paper somewhere with a code that you dial into the phone and press the # then the screen turns on and it shows an easter egg.