Added "Y"
I added the Y texture and put it in the right places. I also changed the "PLAY" to   "PLA" so it looks right when you remove the Y. I think we should maybe have a "ghost Y" or something in the place it is supposed to go, because it is not obvious that there is supposed to be 3 bars in the center of the wheel.