Added a "Ghost Y"
I added a flashing Y to the Handle without Bars. I also made it so that the handle will "jump" if you click it without putting in the Y first. I think that this will make it easier to know what is supposed to happen with the handle. I also made it so the Interactable layer does not collide, so you can fit through the door.