Added the Vault
I added the vault, and then organized and renamed the parts. I also deleted a big folder of all my blender files, which I don't know how it got in there but that is most of the changes. all the models that we need are still in the project and the game works fine.