Added the Hard hat but it does not work yet.
I added the hard hat and then tried to make so you can pick it up. I had to change the Item Holder script so it was compatible with 3d, and now you can pick up the hard hat and place it down, but it does not show up in your inventory. Could you try to fix this?