Implemented 3rd door and fixed some things
Dealt with a a lot of merge conflicts, changed some lighting with the fake screens, made the hole not go offscreen, tried to add HDRP then realized you can't use it with webgl and removed it, fixed some issues with the vault (unity is really weird and mesh colliders need to be convex to collide but if there are holes or inside angles (if it is concave) in a mesh and you make it concave you can't go through the hole and it blocks your ability to pick things up so I fixed some things like that) and made the hardhat able to be picked up (there was an extra space in the name of the item), added the keycard item (placeholder is unpause2), added door 3 (the room looks really good now) but the animations are not working, did some file organization, and fixed the keyhole glitch by disabling the play button when the animation is running (though it may have issues when we make the real pause system). Also can you please try to make the doors closable (that was another one of Adrian's suggestions)?