Fixed Door 3
I fixed door 3 and it will animate now. The only problem is it opens without the keycard, so please make it so you need the keycard to open it. On the closing doors thing, I don't know if we need that, because it is not important to the puzzle. I also think that the keycard should be less obvious, and perhaps we could come up with a way to use words to get the keycard. I was thinking that you could get a key from one of the deposit boxes, and then add that to the card of cardboard on a cardboard box on a table, or some version of this.