Some things
In the last commit, I forgot to mention that I changed the ItemHolder script again to take in a game object as well as a image and meshRenderer. This is needed for items (like the keycard) that have multiple objects. Be carful when using this though, the game object has to be different then the one with the ItemHolder script