Changed the doors
Temporarily disabled the closing door (by swapping the animators) because of the glitching out and made the door3 handle wiggle work. Also, do you think it is a good idea for the next room to have pipes continuing from the office with a leak that you have to fix (there is water blocking the door or something) by taking the phone and using it as a connector of the 2 holes (like a pipe junction)? Also, I think that  there could be a paper with a code in some later room that you could put into the phone (but you could only do it once you had interacted with the paper to prevent cheating) to make a picture of the key to another room  (or just something that you needed in general) that you could take out of the screen and use, so you don't just go forward in the rooms you have to go back sometimes like an escape room. Lastly, I think that we should use the "red" in "credits" somewhere where it would turn into the color red. An example is there would be a security laser that you couldn't get past with a status light that was green but if you put the red on the light it turns red and the lasers shut off so you can go through. Or maybe at the end of the game right before the hat part there is an airlock or something that has a green light for secure and you can turn it red to open the door then do the hat thing and escape. What do you think about these ideas?