Possible keycard location
I put the keycard (which would in the story be a skeleton key) in the slot for the open vault like someone left it there so you can use it for the door. I feel like we should leave as many things as possible in the earlier rooms to go back too but also make the puzzles not too hard so the difficulty can increase later on, but tell me what you think about this.