Some more possible changes
Made the phone an item (placeholder is ghost y) and temporarily disabled the door. Without the door, I feel like the experience of first seeing the office makes you think of the hallway as more of an access tunnel or maintenance area than a security corridor, and also the office looks a lot more impressive for some reason and after the boring gray tiles you feel kind of overwhelmed with all of the new stuff in the room, like the gray box was a facade and you are breaking out, which I think really fits the game. Also, I think it would be better to have the jumping bar changed to something like the keycard door one where you try to spin the handle but you can't. That might make it seem like it's just locked though, so I don't know.