New Door
I added a different door at the end of the second hallway. Perhaps it could be opened by some kind of electrical box on the wall as the walls of the hallway are kind of sparse. For now you can click on the door to open it. Also, what kind of room could go after this?