Many things
I added a new puzzle to open the door. I added a electrical box in the hallway. Click on it and the front will fall down. You have to use the phone to fix the broken wire. I also added a new function to Interactable Object Script that runs a second event. This can be triggered in an animation, so it is helpful for when we need something to happen after an animation.