Fixed Electrical Panel
I fixed the electrical panel, now it takes 3 clicks to open the door, one to open the electrical panel, one to place the phone and one to open the drawer and hit the button. The thoughts are really great. However, there are 2 things I think we should add. One is that the text fades in and out, and also that for the text that appears when you look at things, there should be a range I don't know how you would do this, but it is kind of weird if you are just looking around the room to get a random dialogue.